chore(cli-forge): reorganize some code
