+++ packages/cli-forge/src/index.ts
@@ -1,2 +1,3 @@
-export { default, cli, CLI, Command } from './lib/cli-forge';
 export { TestHarness } from './lib/test-harness';
+export * from './lib/public-api';
+export { default } from './lib/public-api';
