+++ packages/cli-forge/src/lib/cli-forge.spec.ts
@@ -1,4 +1,4 @@
-import { cli } from './cli-forge';
+import { cli } from './public-api';
 
 const ORIGINAL_CONSOLE_LOG = console.log;
